test(sqla repo): unify UUID and BigInt tests (#2093)
* test(sqla repo): unify UUID and BigInt tests

* test(sqla repo): rename test_repo_uuid.py > test_repository.py

* test(sqla repo): fix typing for 3.8

* test(sqla repo): fix typing

* feat: test stringified and raw IDs

* test: address review comments and do some refactoring

* Revert "feat: test stringified and raw IDs"

This reverts commit 512e4540

* test(sqla repo): unify UUID and BigInt tests (Sourcery refactored) (#2108)

'Refactored by Sourcery'


* test: fix typing that sourcery broke

---------
